Organize utils properly
