+++ lib/modules/editor/screens/node.dart
@@ -3,8 +3,8 @@ import 'package:andax/models/cell_write.dart';
 import 'package:andax/models/node.dart';
 import 'package:andax/models/transition.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/modules/editor/services/node.dart';
-import 'package:andax/modules/editor/services/pickers.dart';
+import 'package:andax/modules/editor/utils/node.dart';
+import 'package:andax/modules/editor/utils/pickers.dart';
 import 'package:andax/modules/editor/widgets/transition_dialog.dart';
 import 'package:andax/shared/widgets/options_button.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
