+++ lib/modules/editor/widgets/cell_write_dialog.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../screens/story.dart';
-import '../services/pickers.dart';
+import '../utils/pickers.dart';
 
 Future<CellWrite?> showCellWriteDialog(
   BuildContext context,
