+++ lib/modules/editor/widgets/transition_dialog.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../screens/story.dart';
-import '../services/pickers.dart';
+import '../utils/pickers.dart';
 
 Future<Transition?> showTransitionEditorDialog(
   BuildContext context,
