+++ lib/modules/home/screens/home.dart
@@ -11,7 +11,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
 import '../services/searching.dart';
-import '../services/sheets.dart';
+import '../utils/sheets.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
