+++ lib/modules/home/screens/search.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:algolia/algolia.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/modules/home/services/searching.dart';
-import 'package:andax/modules/home/services/sheets.dart';
+import 'package:andax/modules/home/utils/sheets.dart';
 import 'package:andax/modules/home/widgets/story_tile.dart';
 import 'package:andax/shared/extensions.dart';
 import 'package:andax/shared/widgets/paging_list.dart';
