+++ lib/modules/profile/screens/profile.dart
@@ -1,5 +1,5 @@
 import 'package:andax/models/story.dart';
-import 'package:andax/modules/home/services/sheets.dart';
+import 'package:andax/modules/home/utils/sheets.dart';
 import 'package:andax/modules/home/widgets/story_tile.dart';
 import 'package:andax/shared/widgets/paging_list.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
